+++ frontend/src/app/index.css
@@ -1,0 +1,28 @@
+body {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    min-height: 100vh;
+    text-align: center;
+    margin-bottom: 1rem;
+}
+
+button {
+    padding: 0.75rem;
+    padding-left: 3rem;
+    padding-right: 3rem;
+}
+
+.goodmark {
+    font-size: 2rem;
+}
+
+.goodmark, p {
+    margin-bottom: 1rem;
+}
+
+img {
+    margin: auto;
+    margin-bottom: 1.5rem;
+}
