+++ frontend/src/app/layout.js
@@ -1,11 +1,13 @@
 import "prismjs/themes/prism.css";
 import "./globals.css";
+import Link from "next/link";
 
 
 export const metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "GoodMark",
+  description: "A better way to write Markdown",
 };
+
 
 export default function RootLayout({ children }) {
   return (
